Add files via upload
A keylogger, or keystroke logger, is a type of surveillance software or hardware that records every keystroke made on a computer keyboard. This data can be used to monitor and capture sensitive information such as usernames, passwords, and other confidential data. Keyloggers can be implemented for legitimate purposes, like monitoring employee activity or recovering lost data, but they are often associated with malicious intent, such as stealing personal information or conducting unauthorized surveillance. Due to their potential for misuse, the creation and use of keyloggers are subject to strict legal and ethical guidelines.